Changes to make tests and examples compile

+++ test/unit/ut_listeners.cpp
@@ -67,7 +67,7 @@ public:
     cancel_rejects_.push_back(order);
   }
   virtual void on_replace(const OrderPtr& order,
-                          const int32_t& , // size_delta
+                          const int64_t& , // size_delta
                           Price )          // new_price)
   {
     replaces_.push_back(order);
